+++ src/components/mypage/EditProfileImage.tsx
@@ -1,15 +1,17 @@
 import { supabase } from '@/app/api/supabase/supabase';
 import Image from 'next/image';
-import { ChangeEvent, Dispatch, SetStateAction, useRef } from 'react';
+import { ChangeEvent, Dispatch, SetStateAction, useRef, useState } from 'react';
 import basicProfileImage from '../../../public/profile-image.png';
+import { useStartTyping } from 'react-use';
 
 interface EditProfileImageProps {
   newProfileImage: string;
-  setNewProfileImage: Dispatch<SetStateAction<string>>;
   isEditing: boolean;
+  selectedImage: File | null;
+  setSelectedImage: React.Dispatch<React.SetStateAction<File>>;
 }
 
-const EditProfileImage = ({ newProfileImage, setNewProfileImage, isEditing }: EditProfileImageProps) => {
+const EditProfileImage = ({ newProfileImage, isEditing, selectedImage, setSelectedImage }: EditProfileImageProps) => {
   const fileInput = useRef<HTMLInputElement | null>(null);
 
   // 프로필 이미지 수정 버튼 클릭
@@ -17,33 +19,19 @@ const EditProfileImage = ({ newProfileImage, setNewProfileImage, isEditing }: Ed
     fileInput.current?.click();
   };
 
-  // supabase storage에 프로필 이미지 업로드
-  const uploadProfileImage = async (file: File) => {
-    const randomUUID = crypto.randomUUID();
-    const filePath = `profile/${randomUUID}`;
-    const { data, error } = await supabase.storage.from('profileImages').upload(filePath, file);
-    if (error) {
-      console.error('파일 업로드 실패 :', error);
-      throw error;
-    } else {
-      const url = `${process.env.NEXT_PUBLIC_SUPABASE_URL}/storage/v1/object/public/profileImages/${data.path}`;
-      // console.log('url', url);
-      return setNewProfileImage(url);
-    }
-  };
-
   // 수정된 프로필 이미지 반영
   const handleOnChangeImage = (e: ChangeEvent<HTMLInputElement>) => {
     const file: File | undefined = e.target.files?.[0];
-    if (!file) {
+    // console.log('file', file);
+    if (file) {
+      setSelectedImage(file);
+    } else {
       return;
     }
-
-    uploadProfileImage(file);
   };
 
   // 프로필 이미지가 없을 때, 기본 프로필 이미지 보여주기
-  const profileImage = newProfileImage ? newProfileImage : basicProfileImage;
+  const profileImage = selectedImage ? URL.createObjectURL(selectedImage) : newProfileImage || basicProfileImage;
 
   return (
     <div>
